Crate tentacle

A multiplexed p2p network framework based on yamux that supports mounting custom protocols.

The crate is aimed at implementing a framework that light weight, simple, reliable, high performance, and friendly to users.

Concept

Multiaddr

Multiaddr aims to make network addresses future-proof, composable, and efficient.

It can express almost all network protocols, such as:

TCP/IP: /ip4/127.0.0.1/tcp/1337

DNS/IP: /dns4/localhost/tcp/1337

UDP: /ip4/127.0.0.1/udp/1234

Re-exports

pub use bytes;

pub use multiaddr;

pub use secio;

pub use yamux;

Modules

builder
Some gadgets that help create a service

context
Context for Session and Service

error
Error

protocol_select
Protocol select

service
An abstraction of p2p service

traits
Useful traits

utils
Some useful functions

Structs

ProtocolId
Protocol id

SessionId
Index of session