Struct tentacle::builder::MetaBuilder

pub struct MetaBuilder { /* fields omitted */ }

Builder for protocol meta

Methods

impl MetaBuilder

pub fn new() -> Self

New a default builder

pub fn id(self, id: ProtocolId) -> Self

Define protocol id

It is just an internal index of the system that identifies the open/close and message transfer for the specified protocol.

pub fn name<T: Fn(ProtocolId) -> String + 'static + Send + Sync>(
self,
name: T
) -> Self

Define protocol name, default is "/p2p/protocol_id"

Used to interact with the remote service to determine whether the protocol is supported.

If not found, the protocol connection(not session just sub stream) will be closed, and return a ProtocolSelectError event.

pub fn support_versions(self, versions: Vec<String>) -> Self

Define protocol support versions, default is vec!["0.0.1".to_owned()]

Used to interact with the remote service to confirm that both parties open the same version of the protocol.

If not found, the protocol connection(not session just sub stream) will be closed, and return a ProtocolSelectError event.

pub fn codec<T: Fn() -> Box<dyn Codec + Send + 'static> + 'static + Send + Sync>(
self,
codec: T
) -> Self

Define protocol codec, default is LengthDelimitedCodec

pub fn service_handle<T: FnOnce() -> ProtocolHandle<Box<dyn ServiceProtocol + Send + Unpin + 'static>>>(
self,
service_handle: T
) -> Self

Define protocol service handle, default is neither

pub fn session_handle<T: FnMut() -> ProtocolHandle<Box<dyn SessionProtocol + Send + Unpin + 'static>> + Send + 'static>(
self,
session_handle: T
) -> Self

Define protocol session handle, default is neither

pub fn select_version<T>(self, f: T) -> Self where
T: Fn() -> Option<SelectFn<String>> + Send + Sync + 'static,

Protocol version selection rule, default is select_version

pub fn before_send<T>(self, f: T) -> Self where
T: Fn(Bytes) -> Bytes + 'static + Send,

Unified processing of messages before they are sent

pub fn before_receive<T>(self, f: T) -> Self where
T: Fn() -> Option<Box<dyn Fn(BytesMut) -> Result<Bytes, Error> + Send + 'static>> + Send + Sync + 'static,

Unified processing of messages before user received

pub fn flag(self, flag: BlockingFlag) -> Self

Set a flag to control function behavior

pub fn build(self) -> ProtocolMeta

Combine the configuration of this builder to create a ProtocolMeta