Struct tentacle::context::ServiceContext

pub struct ServiceContext { /* fields omitted */ }

The Service runtime can send some instructions to the inside of the handle. This is the sending channel.

Methods

impl ServiceContext

pub fn listen(&self, address: Multiaddr) -> Result<(), SendErrorKind>

Create a new listener

pub fn dial(
&self,
address: Multiaddr,
target: TargetProtocol
) -> Result<(), SendErrorKind>

Initiate a connection request to address

pub fn disconnect(&self, session_id: SessionId) -> Result<(), SendErrorKind>

Disconnect a connection

pub fn send_message_to(
&self,
session_id: SessionId,
proto_id: ProtocolId,
data: Bytes
) -> Result<(), SendErrorKind>

Send message

pub fn quick_send_message_to(
&self,
session_id: SessionId,
proto_id: ProtocolId,
data: Bytes
) -> Result<(), SendErrorKind>

Send message on quick channel

pub fn filter_broadcast(
&self,
session_ids: TargetSession,
proto_id: ProtocolId,
data: Bytes
) -> Result<(), SendErrorKind>

Send data to the specified protocol for the specified sessions.

pub fn quick_filter_broadcast(
&self,
session_ids: TargetSession,
proto_id: ProtocolId,
data: Bytes
) -> Result<(), SendErrorKind>

Send data to the specified protocol for the specified sessions on quick channel.

pub fn future_task<T>(&self, task: T) -> Result<(), SendErrorKind> where
T: Future<Output = ()> + 'static + Send,

Send a future task

pub fn open_protocol(
&self,
session_id: SessionId,
proto_id: ProtocolId
) -> Result<(), SendErrorKind>

Try open a protocol

If the protocol has been open, do nothing

pub fn open_protocols(
&self,
session_id: SessionId,
target: TargetProtocol
) -> Result<(), SendErrorKind>

Try open protocol

If the protocol has been open, do nothing

pub fn close_protocol(
&self,
session_id: SessionId,
proto_id: ProtocolId
) -> Result<(), SendErrorKind>

Try close a protocol

If the protocol has been closed, do nothing

pub fn control(&self) -> &ServiceControl

Get the internal channel sender side handle

pub fn protocols(&self) -> &Arc<HashMap<ProtocolId, ProtocolInfo>>

Get service protocol message, Map(ID, Name), but can't modify

pub fn key_pair(&self) -> Option<&SecioKeyPair>

Get the key pair of self

pub fn listens(&self) -> &[Multiaddr]

Get service listen address list

pub fn set_service_notify(
&self,
proto_id: ProtocolId,
interval: Duration,
token: u64
) -> Result<(), SendErrorKind>

Set a service notify token

pub fn set_session_notify(
&self,
session_id: SessionId,
proto_id: ProtocolId,
interval: Duration,
token: u64
) -> Result<(), SendErrorKind>

Set a session notify token

pub fn remove_service_notify(
&self,
proto_id: ProtocolId,
token: u64
) -> Result<(), SendErrorKind>

Remove a service timer by a token

pub fn remove_session_notify(
&self,
session_id: SessionId,
proto_id: ProtocolId,
token: u64
) -> Result<(), SendErrorKind>

Remove a session timer by a token

pub fn close(&self) -> Result<(), SendErrorKind>

Close service.

Order:

close all listens

try close all session's protocol stream

try close all session

close service

pub fn shutdown(&self) -> Result<(), SendErrorKind>

Shutdown service, don't care anything, may cause partial message loss