Module tentacle::service

An abstraction of p2p service

Structs

BlockingFlag
Control whether the protocol handle method requires blocking to run default is 0b1111, all function use blocking

ProtocolMeta
Define the minimum data required for a custom protocol

Service
An abstraction of p2p service, currently only supports TCP protocol

ServiceControl
Service control, used to send commands externally at runtime

Enums

ProtocolEvent
Event generated by all protocol

ProtocolHandle
Protocol handle Contains four modes, each of which has a corresponding behavior, please carefully consider which mode should be used in the protocol

ServiceError
Error generated by the Service

ServiceEvent
Event generated by the Service

SessionType
Indicates the session type

TargetProtocol
When dial, specify which protocol want to open

TargetSession
When sending a message, select the specified session