Enum tentacle::service::ServiceError

pub enum ServiceError {
    DialerError {
        address: Multiaddr,
        error: DialerErrorKind,
    },
    ListenError {
        address: Multiaddr,
        error: ListenErrorKind,
    },
    ProtocolSelectError {
        proto_name: Option<String>,
        session_context: Arc<SessionContext>,
    },
    ProtocolError {
        id: SessionId,
        proto_id: ProtocolId,
        error: Error,
    },
    SessionTimeout {
        session_context: Arc<SessionContext>,
    },
    MuxerError {
        session_context: Arc<SessionContext>,
        error: Error,
    },
    ProtocolHandleError {
        proto_id: ProtocolId,
        error: ProtocolHandleErrorKind,
    },
    SessionBlocked {
        session_context: Arc<SessionContext>,
    },
}

Error generated by the Service

Variants

DialerError

When dial remote error

Fields of DialerError

address: Multiaddr

Remote address

error: DialerErrorKind

error

ListenError

When listen error

Fields of ListenError

address: Multiaddr

Listen address

error: ListenErrorKind

error

ProtocolSelectError

Protocol select fail

Fields of ProtocolSelectError

proto_name: Option<String>

Protocol name, if none, timeout or other net problem, if Some, don't support this proto

session_context: Arc<SessionContext>

Session context

ProtocolError

Protocol error during interaction

Fields of ProtocolError

id: SessionId

Session id

proto_id: ProtocolId

Protocol id

error: Error

Codec error

SessionTimeout

After initializing the connection, the session does not open any protocol, suspected fd attack

Fields of SessionTimeout

session_context: Arc<SessionContext>

Session context

MuxerError

Multiplex protocol error

Fields of MuxerError

session_context: Arc<SessionContext>

Session context

error: Error

error, such as InvalidData

ProtocolHandleError

Protocol handle error, will cause memory leaks/abnormal CPU usage

Fields of ProtocolHandleError

proto_id: ProtocolId

Protocol id

error: ProtocolHandleErrorKind

error

SessionBlocked

Session blocked, can't send message, may blocking global system, If the task is too heavy in a short time, it may be repeated multiple times.

Fields of SessionBlocked

session_context: Arc<SessionContext>

Session context