Struct tentacle::service::ProtocolMeta

pub struct ProtocolMeta { /* fields omitted */ }

Define the minimum data required for a custom protocol

Methods

impl ProtocolMeta

pub fn id(&self) -> ProtocolId

Protocol id

pub fn name(&self) -> String

Protocol name, default is "/p2p/protocol_id"

pub fn support_versions(&self) -> Vec<String>

Protocol supported version

pub fn codec(&self) -> Box<dyn Codec + Send + 'static>

The codec used by the custom protocol, such as LengthDelimitedCodec by tokio

pub fn service_handle(
&mut self
) -> ProtocolHandle<Box<dyn ServiceProtocol + Send + Unpin + 'static>>

A service level callback handle for a protocol.

Behavior

This function is called when the protocol is first opened in the service and remains in memory until the entire service is closed.

Warning

Only can be called once, and will return ProtocolHandle::Neither or later.

pub fn session_handle(
&mut self
) -> ProtocolHandle<Box<dyn SessionProtocol + Send + Unpin + 'static>>

A session level callback handle for a protocol.

Behavior

When a session is opened, whenever the protocol of the session is opened, the function will be called again to generate the corresponding exclusive handle.

Correspondingly, whenever the protocol is closed, the corresponding exclusive handle is cleared.

pub fn blocking_flag(&self) -> BlockingFlag

Control whether the protocol handle method requires blocking to run